Fix error: function called through a non-compatible type [-Werror]
Fix IsBallista

+++ include/bmarch.h
@@ -7,7 +7,7 @@ int GetSomeBallistaItemAt(int xPos, int yPos);
 struct Trap* AddBallista(int xPos, int yPos, int ballistaType);
 void RideBallista(struct Unit* unit);
 void TryRemoveUnitFromBallista(struct Unit* unit);
-s8 IsBallista(struct Trap* trap);
+s32 IsBallista(struct Trap* trap);
 int sub_8037AC0(struct Trap* trap);
 int sub_8037AEC(struct Trap* trap);
 int GetBallistaItemUses(struct Trap* trap);
